+++ src/main/java/net/instantgratification/mcainclusive/mixin/CommonVillagerInterfaceMixin.java
@@ -159,7 +159,7 @@ public interface CommonVillagerInterfaceMixin {
                         // 7. Translate back from pivot
                         matrices.translate(-leftPivotX, -leftPivotY, -leftPivotZ);
 
-                        VertexConsumer clippingVertices = new net.instantgratification.mcainclusive.render.ClippingVertexConsumer(vertices, -1.5f / 16.0f);
+                        VertexConsumer clippingVertices = new net.instantgratification.mcainclusive.render.ClippingVertexConsumer(vertices, matrices.last().pose(), -2.0f / 16.0f);
                         cubes.get(0).compile(matrices.last(), clippingVertices, light, overlay, color);
                         matrices.popPose();
                     }
@@ -198,7 +198,7 @@ public interface CommonVillagerInterfaceMixin {
                         // 7. Translate back from pivot
                         matrices.translate(-rightPivotX, -rightPivotY, -rightPivotZ);
 
-                        VertexConsumer clippingVertices = new net.instantgratification.mcainclusive.render.ClippingVertexConsumer(vertices, -1.5f / 16.0f);
+                        VertexConsumer clippingVertices = new net.instantgratification.mcainclusive.render.ClippingVertexConsumer(vertices, matrices.last().pose(), -2.0f / 16.0f);
                         cubes.get(1).compile(matrices.last(), clippingVertices, light, overlay, color);
                         matrices.popPose();
                     }